requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

grpcio 
setuptools 
azure-functions
implicit
numpy
scipy
pandas
scikit-learn
matplotlib
seaborn